+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2018-12-21
+### Fixed
+- Fix autovote not voting
+
+## [0.1.9] - 2018-12-10
+### Fixed
+- Fix multiple errors with Dubtrack changes
+
+### Removed
+- Removed perfect-scrollbar support
+
+## [0.1.8] - 2018-05-22
+### Added
+- Add support for FrankerFaceZ emotes
+
+### Changed
+- Update custom mentions to only trigger when a whole word is matched rather than a partial word
+
 ## [0.1.7] - 2017-11-22
 ### Changed
 - Update Twitch emojis to save to IndexedDB instead of localStorage. We were hitting the 5mb storage limit on localStorage, and using IndexedDB increases the limit to 50mb.